move process specific variables to the corresponding process
