+++ dsf-bpe/dsf-bpe-process-base/src/main/java/org/highmed/dsf/fhir/variables/SerializerPlugin.java
@@ -1,7 +1,11 @@
 package org.highmed.dsf.fhir.variables;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.camunda.bpm.engine.ProcessEngine;
 import org.camunda.bpm.engine.impl.cfg.ProcessEngineConfigurationImpl;
@@ -11,29 +15,28 @@ import org.camunda.bpm.engine.impl.variable.serializer.TypedValueSerializer;
 public class SerializerPlugin implements ProcessEnginePlugin
 {
 	@SuppressWarnings("rawtypes")
-	private final List<TypedValueSerializer> serializer;
+	private List<TypedValueSerializer> serializer;
 
 	public SerializerPlugin(FhirResourceSerializer fhirResourceSerializer,
 			FhirResourcesListSerializer fhirResourcesListSerializer,
 			MultiInstanceTargetSerializer multiInstanceTargetSerializer,
-			MultiInstanceTargetsSerializer multiInstanceTargetsSerializer,
-			FeasibilityQueryResultSerializer feasibilityQueryResultSerializer,
-			FeasibilityQueryResultsSerializer feasibilityQueryResultsSerializer,
-			FinalFeasibilityQueryResultSerializer finalFeasibilityQueryResultSerializer,
-			FinalFeasibilityQueryResultsSerializer finalFeasibilityQueryResultsSerializer,
-			BloomFilterConfigSerializer bloomFilterConfigSerializer, OutputSerializer outputSerializer,
+			MultiInstanceTargetsSerializer multiInstanceTargetsSerializer, OutputSerializer outputSerializer,
 			OutputsSerializer outputsSerializer)
 	{
 		serializer = Arrays.asList(fhirResourceSerializer, fhirResourcesListSerializer, multiInstanceTargetSerializer,
-				multiInstanceTargetsSerializer, feasibilityQueryResultSerializer, feasibilityQueryResultsSerializer,
-				finalFeasibilityQueryResultSerializer, finalFeasibilityQueryResultsSerializer,
-				bloomFilterConfigSerializer, outputSerializer, outputsSerializer);
+				multiInstanceTargetsSerializer, outputSerializer, outputsSerializer);
 	}
 
 	@Override
 	public void preInit(ProcessEngineConfigurationImpl processEngineConfiguration)
 	{
-		processEngineConfiguration.setCustomPreVariableSerializers(serializer);
+		List<TypedValueSerializer> initialSerializers = Optional.ofNullable(processEngineConfiguration
+				.getCustomPreVariableSerializers()).orElse(Collections.emptyList());
+
+		List<TypedValueSerializer> allSerializer = Stream.concat(initialSerializers.stream(), serializer.stream())
+				.collect(Collectors.toList());
+
+		processEngineConfiguration.setCustomPreVariableSerializers(allSerializer);
 	}
 
 	@Override
