+++ dsf-bpe/dsf-bpe-process-feasibility/src/main/java/org/highmed/dsf/bpe/message/SendMultiMedicResults.java
@@ -8,8 +8,8 @@ import org.highmed.dsf.fhir.client.FhirWebserviceClientProvider;
 import org.highmed.dsf.fhir.organization.OrganizationProvider;
 import org.highmed.dsf.fhir.task.AbstractTaskMessageSend;
 import org.highmed.dsf.fhir.task.TaskHelper;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResult;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResults;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResult;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResults;
 import org.hl7.fhir.r4.model.Extension;
 import org.hl7.fhir.r4.model.Reference;
 import org.hl7.fhir.r4.model.Task.ParameterComponent;
