+++ dsf-bpe/dsf-bpe-process-feasibility/src/main/java/org/highmed/dsf/bpe/service/CheckMultiMedicResults.java
@@ -9,8 +9,8 @@ import org.highmed.dsf.bpe.delegate.AbstractServiceDelegate;
 import org.highmed.dsf.bpe.variables.ConstantsFeasibility;
 import org.highmed.dsf.fhir.client.FhirWebserviceClientProvider;
 import org.highmed.dsf.fhir.task.TaskHelper;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResult;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResults;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResult;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResults;
 import org.hl7.fhir.r4.model.Extension;
 import org.hl7.fhir.r4.model.Reference;
 import org.hl7.fhir.r4.model.Task;
