+++ dsf-bpe/dsf-bpe-process-feasibility/src/main/java/org/highmed/dsf/bpe/service/ExecuteRecordLink.java
@@ -11,11 +11,11 @@ import org.highmed.dsf.bpe.delegate.AbstractServiceDelegate;
 import org.highmed.dsf.bpe.variables.ConstantsFeasibility;
 import org.highmed.dsf.fhir.client.FhirWebserviceClientProvider;
 import org.highmed.dsf.fhir.task.TaskHelper;
-import org.highmed.dsf.fhir.variables.FeasibilityQueryResult;
-import org.highmed.dsf.fhir.variables.FeasibilityQueryResults;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResult;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResults;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResultsValues;
+import org.highmed.dsf.bpe.variables.FeasibilityQueryResult;
+import org.highmed.dsf.bpe.variables.FeasibilityQueryResults;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResult;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResults;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResultsValues;
 import org.highmed.pseudonymization.domain.PersonWithMdat;
 import org.highmed.pseudonymization.domain.impl.MatchedPersonImpl;
 import org.highmed.pseudonymization.recordlinkage.FederatedMatcherImpl;
