+++ dsf-bpe/dsf-bpe-process-feasibility/src/main/java/org/highmed/dsf/bpe/variables/FeasibilityQueryResultSerializer.java
@@ -1,4 +1,4 @@
-package org.highmed.dsf.fhir.variables;
+package org.highmed.dsf.bpe.variables;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -6,20 +6,20 @@ import java.util.Objects;
 import org.camunda.bpm.engine.impl.variable.serializer.PrimitiveValueSerializer;
 import org.camunda.bpm.engine.impl.variable.serializer.ValueFields;
 import org.camunda.bpm.engine.variable.impl.value.UntypedValueImpl;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResultValues.FinalFeasibilityQueryResultValue;
+import org.highmed.dsf.bpe.variables.FeasibilityQueryResultValues.FeasibilityQueryResultValue;
 import org.springframework.beans.factory.InitializingBean;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-public class FinalFeasibilityQueryResultSerializer extends PrimitiveValueSerializer<FinalFeasibilityQueryResultValue>
+public class FeasibilityQueryResultSerializer extends PrimitiveValueSerializer<FeasibilityQueryResultValue>
 		implements InitializingBean
 {
 	private final ObjectMapper objectMapper;
 
-	public FinalFeasibilityQueryResultSerializer(ObjectMapper objectMapper)
+	public FeasibilityQueryResultSerializer(ObjectMapper objectMapper)
 	{
-		super(FinalFeasibilityQueryResultValues.VALUE_TYPE);
+		super(FeasibilityQueryResultValues.VALUE_TYPE);
 
 		this.objectMapper = objectMapper;
 	}
@@ -31,9 +31,9 @@ public class FinalFeasibilityQueryResultSerializer extends PrimitiveValueSeriali
 	}
 
 	@Override
-	public void writeValue(FinalFeasibilityQueryResultValue value, ValueFields valueFields)
+	public void writeValue(FeasibilityQueryResultValue value, ValueFields valueFields)
 	{
-		FinalFeasibilityQueryResult result = value.getValue();
+		FeasibilityQueryResult result = value.getValue();
 		try
 		{
 			if (result != null)
@@ -46,21 +46,21 @@ public class FinalFeasibilityQueryResultSerializer extends PrimitiveValueSeriali
 	}
 
 	@Override
-	public FinalFeasibilityQueryResultValue convertToTypedValue(UntypedValueImpl untypedValue)
+	public FeasibilityQueryResultValue convertToTypedValue(UntypedValueImpl untypedValue)
 	{
-		return FinalFeasibilityQueryResultValues.create((FinalFeasibilityQueryResult) untypedValue.getValue());
+		return FeasibilityQueryResultValues.create((FeasibilityQueryResult) untypedValue.getValue());
 	}
 
 	@Override
-	public FinalFeasibilityQueryResultValue readValue(ValueFields valueFields, boolean asTransientValue)
+	public FeasibilityQueryResultValue readValue(ValueFields valueFields, boolean asTransientValue)
 	{
 		byte[] bytes = valueFields.getByteArrayValue();
 
 		try
 		{
-			FinalFeasibilityQueryResult result = (bytes == null || bytes.length <= 0) ? null
-					: objectMapper.readValue(bytes, FinalFeasibilityQueryResult.class);
-			return FinalFeasibilityQueryResultValues.create(result);
+			FeasibilityQueryResult result = (bytes == null || bytes.length <= 0) ? null
+					: objectMapper.readValue(bytes, FeasibilityQueryResult.class);
+			return FeasibilityQueryResultValues.create(result);
 		}
 		catch (IOException e)
 		{
