+++ dsf-bpe/dsf-bpe-process-feasibility/src/main/java/org/highmed/dsf/bpe/variables/FeasibilityQueryResults.java
@@ -1,4 +1,4 @@
-package org.highmed.dsf.fhir.variables;
+package org.highmed.dsf.bpe.variables;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -8,30 +8,30 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class FinalFeasibilityQueryResults
+public class FeasibilityQueryResults
 {
-	private final List<FinalFeasibilityQueryResult> results = new ArrayList<>();
+	private final List<FeasibilityQueryResult> results = new ArrayList<>();
 
 	@JsonCreator
-	public FinalFeasibilityQueryResults(@JsonProperty("results") Collection<? extends FinalFeasibilityQueryResult> results)
+	public FeasibilityQueryResults(@JsonProperty("results") Collection<? extends FeasibilityQueryResult> results)
 	{
 		if (results != null)
 			this.results.addAll(results);
 	}
 
-	public void add(FinalFeasibilityQueryResult newResult)
+	public void add(FeasibilityQueryResult newResult)
 	{
 		if (newResult != null)
 			results.add(newResult);
 	}
 
-	public void addAll(Collection<FinalFeasibilityQueryResult> results)
+	public void addAll(Collection<FeasibilityQueryResult> results)
 	{
 		if (results != null)
 			this.results.addAll(results);
 	}
 
-	public List<FinalFeasibilityQueryResult> getResults()
+	public List<FeasibilityQueryResult> getResults()
 	{
 		return Collections.unmodifiableList(results);
 	}
