+++ dsf-bpe/dsf-bpe-process-feasibility/src/main/java/org/highmed/dsf/bpe/variables/FinalFeasibilityQueryResultSerializer.java
@@ -1,4 +1,4 @@
-package org.highmed.dsf.fhir.variables;
+package org.highmed.dsf.bpe.variables;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -6,20 +6,20 @@ import java.util.Objects;
 import org.camunda.bpm.engine.impl.variable.serializer.PrimitiveValueSerializer;
 import org.camunda.bpm.engine.impl.variable.serializer.ValueFields;
 import org.camunda.bpm.engine.variable.impl.value.UntypedValueImpl;
-import org.highmed.dsf.fhir.variables.FeasibilityQueryResultValues.FeasibilityQueryResultValue;
+import org.highmed.dsf.bpe.variables.FinalFeasibilityQueryResultValues.FinalFeasibilityQueryResultValue;
 import org.springframework.beans.factory.InitializingBean;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-public class FeasibilityQueryResultSerializer extends PrimitiveValueSerializer<FeasibilityQueryResultValue>
+public class FinalFeasibilityQueryResultSerializer extends PrimitiveValueSerializer<FinalFeasibilityQueryResultValue>
 		implements InitializingBean
 {
 	private final ObjectMapper objectMapper;
 
-	public FeasibilityQueryResultSerializer(ObjectMapper objectMapper)
+	public FinalFeasibilityQueryResultSerializer(ObjectMapper objectMapper)
 	{
-		super(FeasibilityQueryResultValues.VALUE_TYPE);
+		super(FinalFeasibilityQueryResultValues.VALUE_TYPE);
 
 		this.objectMapper = objectMapper;
 	}
@@ -31,9 +31,9 @@ public class FeasibilityQueryResultSerializer extends PrimitiveValueSerializer<F
 	}
 
 	@Override
-	public void writeValue(FeasibilityQueryResultValue value, ValueFields valueFields)
+	public void writeValue(FinalFeasibilityQueryResultValue value, ValueFields valueFields)
 	{
-		FeasibilityQueryResult result = value.getValue();
+		FinalFeasibilityQueryResult result = value.getValue();
 		try
 		{
 			if (result != null)
@@ -46,21 +46,21 @@ public class FeasibilityQueryResultSerializer extends PrimitiveValueSerializer<F
 	}
 
 	@Override
-	public FeasibilityQueryResultValue convertToTypedValue(UntypedValueImpl untypedValue)
+	public FinalFeasibilityQueryResultValue convertToTypedValue(UntypedValueImpl untypedValue)
 	{
-		return FeasibilityQueryResultValues.create((FeasibilityQueryResult) untypedValue.getValue());
+		return FinalFeasibilityQueryResultValues.create((FinalFeasibilityQueryResult) untypedValue.getValue());
 	}
 
 	@Override
-	public FeasibilityQueryResultValue readValue(ValueFields valueFields, boolean asTransientValue)
+	public FinalFeasibilityQueryResultValue readValue(ValueFields valueFields, boolean asTransientValue)
 	{
 		byte[] bytes = valueFields.getByteArrayValue();
 
 		try
 		{
-			FeasibilityQueryResult result = (bytes == null || bytes.length <= 0) ? null
-					: objectMapper.readValue(bytes, FeasibilityQueryResult.class);
-			return FeasibilityQueryResultValues.create(result);
+			FinalFeasibilityQueryResult result = (bytes == null || bytes.length <= 0) ? null
+					: objectMapper.readValue(bytes, FinalFeasibilityQueryResult.class);
+			return FinalFeasibilityQueryResultValues.create(result);
 		}
 		catch (IOException e)
 		{
