+++ dsf-bpe/dsf-bpe-process-feasibility/src/test/java/org/highmed/dsf/bpe/variables/BloomFilterConfigTest.java
@@ -1,4 +1,4 @@
-package org.highmed.dsf.fhir.variables;
+package org.highmed.dsf.bpe.variables;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
