+++ dsf-bpe/dsf-bpe-process-local-services/src/main/java/org/highmed/dsf/bpe/service/ExtractInputValues.java
@@ -8,11 +8,11 @@ import java.util.stream.Stream;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.highmed.dsf.bpe.ConstantsBase;
 import org.highmed.dsf.bpe.delegate.AbstractServiceDelegate;
+import org.highmed.dsf.bpe.variables.BloomFilterConfig;
+import org.highmed.dsf.bpe.variables.BloomFilterConfigValues;
 import org.highmed.dsf.bpe.variables.ConstantsFeasibility;
 import org.highmed.dsf.fhir.client.FhirWebserviceClientProvider;
 import org.highmed.dsf.fhir.task.TaskHelper;
-import org.highmed.dsf.fhir.variables.BloomFilterConfig;
-import org.highmed.dsf.fhir.variables.BloomFilterConfigValues;
 import org.highmed.dsf.fhir.variables.FhirResourcesListValues;
 import org.hl7.fhir.r4.model.Expression;
 import org.hl7.fhir.r4.model.Group;
