+++ dsf-bpe/dsf-bpe-server/src/main/java/org/highmed/dsf/bpe/spring/config/SerializerConfig.java
@@ -1,15 +1,10 @@
 package org.highmed.dsf.bpe.spring.config;
 
 import org.camunda.bpm.engine.impl.cfg.ProcessEnginePlugin;
-import org.highmed.dsf.fhir.variables.BloomFilterConfigSerializer;
-import org.highmed.dsf.fhir.variables.FeasibilityQueryResultSerializer;
-import org.highmed.dsf.fhir.variables.FeasibilityQueryResultsSerializer;
 import org.highmed.dsf.fhir.variables.FhirResourceJacksonDeserializer;
 import org.highmed.dsf.fhir.variables.FhirResourceJacksonSerializer;
 import org.highmed.dsf.fhir.variables.FhirResourceSerializer;
 import org.highmed.dsf.fhir.variables.FhirResourcesListSerializer;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResultSerializer;
-import org.highmed.dsf.fhir.variables.FinalFeasibilityQueryResultsSerializer;
 import org.highmed.dsf.fhir.variables.MultiInstanceTargetSerializer;
 import org.highmed.dsf.fhir.variables.MultiInstanceTargetsSerializer;
 import org.highmed.dsf.fhir.variables.OutputSerializer;
@@ -88,42 +83,10 @@ public class SerializerConfig
 	}
 
 	@Bean
-	public FeasibilityQueryResultSerializer feasibilityQueryResultSerializer()
-	{
-		return new FeasibilityQueryResultSerializer(objectMapper());
-	}
-
-	@Bean
-	public FeasibilityQueryResultsSerializer feasibilityQueryResultsSerializer()
-	{
-		return new FeasibilityQueryResultsSerializer(objectMapper());
-	}
-
-	@Bean
-	public FinalFeasibilityQueryResultSerializer finalFeasibilityQueryResultSerializer()
-	{
-		return new FinalFeasibilityQueryResultSerializer(objectMapper());
-	}
-
-	@Bean
-	public FinalFeasibilityQueryResultsSerializer finalFeasibilityQueryResultsSerializer()
-	{
-		return new FinalFeasibilityQueryResultsSerializer(objectMapper());
-	}
-
-	@Bean
-	public BloomFilterConfigSerializer bloomFilterConfigSerializer()
-	{
-		return new BloomFilterConfigSerializer(objectMapper());
-	}
-
-	@Bean
 	public ProcessEnginePlugin serializerPlugin()
 	{
 		return new SerializerPlugin(fhirResourceSerializer(), fhirResourcesListSerializer(),
-				multiInstanceTargetSerializer(), multiInstanceTargetsSerializer(), feasibilityQueryResultSerializer(),
-				feasibilityQueryResultsSerializer(), finalFeasibilityQueryResultSerializer(),
-				finalFeasibilityQueryResultsSerializer(), bloomFilterConfigSerializer(), outputSerializer(),
+				multiInstanceTargetSerializer(), multiInstanceTargetsSerializer(), outputSerializer(),
 				outputsSerializer());
 	}
 }
